feat(library): move Presets carousel to Your Library page
